+++ mayan/apps/converter/__init__.py
@@ -1,6 +1,5 @@
 from __future__ import unicode_literals
 
-from .runtime import converter_class  # NOQA
 from .transformations import (  # NOQA
     BaseTransformation, TransformationResize, TransformationRotate,
     TransformationZoom